Create Parsing file (manual input inside code)
For future will have to check to make sure there is no other "origin" in the file you are parsing. Right now I have it so that it scans through to make sure there is no other origin but it does not do anything if there is.
Also, this code goes through all cases for inside a sequence (the first sequence) and makes every subsequence possible.
NOTE: as of right now this prints out the chars of the list to be able to test functionality, now for each case we just need to make it so each data value turns into a 2 bit number and stored it into a long.  This shouldn't be too hard especially with the code working as is right now.